+++ passes/S020/testdata/src/a/comment_ignore.go
@@ -1,0 +1,13 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S020
+	_ = schema.Schema{
+		Computed: true,
+		ForceNew: true,
+	}
+}